Inlezen config

+++ ADS_Simulation/Events/Event.cs
@@ -8,7 +8,7 @@ namespace ADS_Simulation.Events
 {
     abstract class Event : StablePriorityQueueNode
     {
-        public ulong time;
+        public int time;
 
         /// <summary>
         /// Execute the event and add new events to the queue
